add NonIterableTypeTrait
