+++ src/Type/JustNullableTypeTrait.php
@@ -46,19 +46,4 @@ trait JustNullableTypeTrait
 		return true;
 	}
 
-	public function isIterable(): TrinaryLogic
-	{
-		return TrinaryLogic::createNo();
-	}
-
-	public function getIterableKeyType(): Type
-	{
-		return new ErrorType();
-	}
-
-	public function getIterableValueType(): Type
-	{
-		return new ErrorType();
-	}
-
 }
