+++ src/Type/StringType.php
@@ -3,32 +3,19 @@
 namespace PHPStan\Type;
 
 use PHPStan\TrinaryLogic;
+use PHPStan\Type\Traits\NonIterableTypeTrait;
 use PHPStan\Type\Traits\NonObjectTypeTrait;
 
 class StringType implements Type
 {
 
 	use JustNullableTypeTrait;
+	use NonIterableTypeTrait;
 	use NonObjectTypeTrait;
 
 	public function describe(): string
 	{
 		return 'string';
-	}
-
-	public function isIterable(): TrinaryLogic
-	{
-		return TrinaryLogic::createNo();
-	}
-
-	public function getIterableKeyType(): Type
-	{
-		return new ErrorType();
-	}
-
-	public function getIterableValueType(): Type
-	{
-		return new ErrorType();
 	}
 
 	public function isCallable(): TrinaryLogic
