+++ src/Type/Traits/NonIterableTypeTrait.php
@@ -1,0 +1,27 @@
+<?php declare(strict_types = 1);
+
+namespace PHPStan\Type\Traits;
+
+use PHPStan\TrinaryLogic;
+use PHPStan\Type\ErrorType;
+use PHPStan\Type\Type;
+
+trait NonIterableTypeTrait
+{
+
+	public function isIterable(): TrinaryLogic
+	{
+		return TrinaryLogic::createNo();
+	}
+
+	public function getIterableKeyType(): Type
+	{
+		return new ErrorType();
+	}
+
+	public function getIterableValueType(): Type
+	{
+		return new ErrorType();
+	}
+
+}
